code restructing and minor fixes
